Replace map of doubles with list of ints

+++ src/main/kotlin/org/incava/rando/RandCalcList.kt
@@ -1,7 +1,7 @@
 package org.incava.rando
 
 class RandCalcList(size: Int, numSlots: Int, numIterations: Int) : RandCalc(size, numSlots, numIterations) {
-    private val list: List<Int> = slots.map { it.value.toInt() }
+    private val list: List<Int> = slots.map { it.value }
 
     override fun nextInt(): Int {
         val index = random.nextInt(numSlots)
